yürüme animasyonu enemy class'ına entegre edildi. İsimlendirmeler standartlaştırıldı.

+++ src/main/java/com/example/termproject2/Infantry.java
@@ -2,12 +2,12 @@ package com.example.termproject2;
 
 public class Infantry extends Enemy
 {
-    Infantry(int initialHealth, int initialSpeed)
+    Infantry(int initialHealth, double initialSpeed)
     {
         super(initialHealth, initialSpeed);
     }
 
-    Infantry(String imagePath, int initialHealth, int initialSpeed)
+    Infantry(String imagePath, int initialHealth, double initialSpeed)
     {
         super(imagePath, initialHealth, initialSpeed);
     }
